+++ src/utils/Collector.java
@@ -2,5 +2,5 @@ package utils;
 
 public interface Collector {
 	public void addEntry(int x);
-	public void addEntry(int x, double va);
+	public void addEntry(int x, float va);
 }
